Allow to change TableContainer description tag name
Description is a useful part of TableContainer's header section, yet its current implementation using <p> tag limits the possibilities of its children. Children of type <div> currently cause invalid DOM nesting since <div> cannot be a descendant of <p>.

This PR keeps the default implementation of description as tag <p>, but allows for an additional prop to be passed to TableContainer to change the tag name of description to be div.